% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addConstraint}
\alias{addConstraint}
\title{Add constraints to model}
\usage{
addConstraint(model, react, coeff, rtype, lb = NULL, ub = NULL)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{Character vector or a list of character vectors containing the
model's reactions IDs that are part of the respective constraint.}

\item{coeff}{Numeric vector or list of numeric vectors defining the
coefficients for the reactions listed in 'react'.}

\item{rtype}{Character vector describing the type of the constraint(s). See
details.}

\item{lb, ub}{Numeric vector defining the lower and upper bound(s) of the
constraint(s).}
}
\description{
Add linear reaction flux constraints to a metabolic network.
}
\details{
The slot "rtype" describes the type of each constraint. Valid values and
their effects are:\tabular{clc}{
   \emph{code} \tab \emph{description} \tab \emph{rule} \cr
   "F" \tab free constraint \tab \eqn{-\infty < x < \infty} \cr
   "L" \tab constraint with lower bound \tab \eqn{lb \leq x \leq \infty} \cr
   "U" \tab constraint with upper bound \tab \eqn{-\infty \leq x \leq ub} \cr
   "D" \tab double-bounded (ranged) constraint \tab \eqn{lb \leq x \leq ub} \cr
   "E" \tab fixed (equality constraint) \tab \eqn{lb = x = ub} \cr
}
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)

# Simulate anaerobic growth
mod <- changeBounds(mod, "EX_o2_e", lb = 0)

# Limit the proton production depending on the growth rate
mod <- addConstraint(mod,
                     react = c("EX_h_e","BIOMASS_Ecoli_core_w_GAM"),
                     coeff = c(1, -20), rtype = "U", ub = 0)

}
