% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addGene}
\alias{addGene}
\title{Add genes or update their data}
\usage{
addGene(
  model,
  id,
  name = NA,
  CVTerms = NA,
  SBOTerm = rep("SBO:0000243", length(id))
)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{id}{Character vector with gene IDs}

\item{name}{Character vector for gene names}

\item{CVTerms}{Character vector for the genes' CV-Terms}

\item{SBOTerm}{Character vector for the genes' SBO-Terms}
}
\description{
The functions allows you to add one or more genes to a model. When
providing the ID of an already existing genes, you can use this function
to update the genes' information.
}
