% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addMetabolite}
\alias{addMetabolite}
\title{Add metabolites or update their data}
\usage{
addMetabolite(
  model,
  id,
  name = NA,
  comp = NA,
  chemicalFormula = NA,
  charge = NA,
  annotation = NA
)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
The functions allows you to add one or more metabolites to a model. When
providing the ID of an already existing metabolite, you can use this function
to update metabolite information.
}
