% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addReact}
\alias{addReact}
\title{Add or modify a reaction}
\usage{
addReact(
  model,
  id,
  met,
  Scoef,
  reversible = FALSE,
  lb = 0,
  ub = COBRAR_SETTINGS("MAXIMUM"),
  obj = 0,
  subSystem = NA,
  subSystemName = NA,
  gprAssoc = NA,
  reactName = NA,
  metName = NA,
  metComp = NA,
  metCharge = NA,
  metChemicalFormula = NA,
  annotation = NA
)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{id}{Character for the reaction ID}

\item{met}{Character vector providing the IDs of metabolites that participate
in the reaction}

\item{Scoef}{Numeric vector (same length as \code{met}) of stoichiometric
coefficients for the metabolites in \code{met}. The value in \code{Scoef[i]} is the
stoichiometric coefficient of the metabolite in \code{met[i]}.}

\item{reversible}{This option has now effect and is only here for legacy
reasons. Whether a reaction is reversible or not is inferred by cobrar based
on the lower and upper bounds.}

\item{lb, ub}{Single numeric values that define the lower and upper flux
limits, respectively.}

\item{obj}{Single numeric value for the coefficient of the reaction in the
objective function.}

\item{subSystem}{A vector of character strings containing the sub system IDs
to which the reaction belongs.}

\item{subSystemName}{A character vector (same length as \code{subSystem}) for the
names of the subsystems. If the subsystem is already part of the model and
you do not want to change its name, just use NA the corresponding entry.}

\item{gprAssoc}{A single character string giving the Gene-Product-Reaction
(GPR) association for the reaction. If NA, no GRP association is created.}

\item{reactName}{A single character string giving the name for the reaction.
If NA, the value of argument \code{id} is used.}

\item{metName}{A vector of character strings of the same length as \code{met}
containing the metabolites names for the metabolites given in argument \code{met}.}

\item{metCharge}{A numeric vector of the same length as \code{met} defining the
charges for the metabolites given in argument \code{met}.}

\item{metChemicalFormula}{A character vector of the same length as \code{met}
defining the chemical formulas for the metabolites given in argument \code{met}.}

\item{annotation}{An annotation string for the reaction.}
}
\description{
The function can be used to add or modify a reaction in an existing model.
}
\details{
If you want to use the function to update data of a pre-existing reaction but
not its stoichiometry, use NA for the parameters 'met' and 'Scoeff'.
If the reaction is already part of the model, any reaction value (e.g., lb,
ub, reactName), that is set to NA has the effect that the old value will be
used.
If the reaction is already part of the model, and values for the parameter
\code{subSystem} are provided, all previous set Subsystem associations of the
reaction will be removed.
If metabolites or subsystems are not part of the model yet, they will be
added.
}
