% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{changeObjFunc}
\alias{changeObjFunc}
\title{Change the objective function}
\usage{
changeObjFunc(model, react, obj_coef = rep(1, length(react)))
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{react}{Character vector containing the model's reactions IDs, that are
part of the new objective function}

\item{obj_coef}{Numeric vector with the objective coefficients of the
reactions in the vector 'react'.}
}
\value{
An updated model of class \link{ModelOrg}
}
\description{
Changes the objective function of a model.
}
\details{
Reactions not listed in 'react' are assigned an objective coefficient of
zero.
}
