% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkGeneId}
\alias{checkGeneId}
\title{Check gene IDs and Indexes}
\usage{
checkGeneId(model, gene)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{gene}{A character vector specifying the gene IDs or a integer
vector providing the gene indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether gene IDs or indexes are part of (valid) for a specific
model.
}
