% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkReactId}
\alias{checkReactId}
\title{Check reaction IDs and Indices}
\usage{
checkReactId(model, react)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{react}{A character vector specifying the reaction IDs or a integer
vector providing the reaction indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether reaction IDs or indices are part of (valid) for a specific
model.
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{geneDel}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
