% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobrar.R
\name{COBRAR_SETTINGS}
\alias{COBRAR_SETTINGS}
\title{Set and get central cobrar parameters}
\usage{
COBRAR_SETTINGS(parm, value, ...)
}
\arguments{
\item{parm}{A character string giving the name of the parameter to set.}

\item{value}{Choose a value to over-write the current parameter.}

\item{...}{TBD.}
}
\description{
Manage a set of default parameter settings for cobrar.
}
\concept{Setup}
