% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{comp_num}
\alias{comp_num}
\alias{comp_num,ModelOrg-method}
\alias{comp_num,ModelOrg}
\title{Number of compartments}
\usage{
comp_num(model)

\S4method{comp_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of compartments of a model
}
