% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{comp_num}
\alias{comp_num}
\title{Number of compartments}
\usage{
comp_num(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of compartments of a model
}
