% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{comp_pos}
\alias{comp_pos}
\alias{comp_pos,modelorg,character-method}
\alias{comp_pos,modelorg,character}
\alias{comp_pos,modelorg,numeric-method}
\alias{comp_pos,modelorg,numeric}
\alias{comp_pos,modelorg,missing-method}
\alias{comp_pos,modelorg,missing}
\alias{comp_pos,modelorg,logical-method}
\alias{comp_pos,modelorg,logical}
\title{Index of compartment(s)}
\usage{
comp_pos(model, comp)

\S4method{comp_pos}{modelorg,character}(model, comp)

\S4method{comp_pos}{modelorg,numeric}(model, comp)

\S4method{comp_pos}{modelorg,missing}(model, comp)

\S4method{comp_pos}{modelorg,logical}(model, comp)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{comp}{Character vector with compartment IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific compartment(s).
}
\details{
Returns NA for compartment IDs not part of the model or if the index is larger
than the number of compartments in the model.
}
