% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{comp_pos}
\alias{comp_pos}
\title{Index of compartment(s)}
\usage{
comp_pos(model, comp)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{gene}{Character vector with compartment IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific compartment(s).
}
\details{
Returns NA for compartment IDs not part of the model or if the index is larger
than the number of compartments in the model.
}
