% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{constraint_num}
\alias{constraint_num}
\alias{constraint_num,ModelOrg-method}
\alias{constraint_num,ModelOrg}
\title{Number of constraints}
\usage{
constraint_num(model)

\S4method{constraint_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of constraints of a model
}
