% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{constraint_num}
\alias{constraint_num}
\title{Number of constraints}
\usage{
constraint_num(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of constraints of a model
}
