% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{countElements}
\alias{countElements}
\title{Count elements in formulas}
\usage{
countElements(formulas, multiplier = 1)
}
\arguments{
\item{formulas}{Character vector of chemical formulas.}

\item{multiplier}{Multiplier for the element counts in each molecule}
}
\value{
A numeric matrix with named columns for individual elements. Rows
contain the counts of elements for each formula in the same order as formulas
were provided to the function.
}
\description{
Counts the number of elements in molecules based on their chemical formulas.
}
\details{
Formulas may not have element numbers containing decimal points.
}
\examples{
countElements(c("C6H12O6","C48H72CoN11O8","HCOOH"))

}
