% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadEndMetabolites.R
\name{deadEndMetabolites}
\alias{deadEndMetabolites}
\title{Identify dead end metabolites}
\usage{
deadEndMetabolites(object)
}
\arguments{
\item{object}{Model of class \link{ModelOrg}}
}
\value{
A list with two elements: "dem" is a character vector with the IDs
of identified dead end metabolites. "der" is a character vector with IDs of
reactions, that involve at least one of the dead end metabolites.
}
\description{
Searches a metabolic network for metabolites that can be produced but not
consumed, and vice versa.
}
\note{
The algorithm is adapted from the original 'sybil' package, which in turn is
adapted from the implementation in 'cobratoolbox'. Important detail: The
algorithm function in the 'sybil' package may return different results than
this cobrar implementation. This is because the 'sybil' function does not
take into account the possibility that an irreversible reaction is defined in
the direction RHS to LHS (lower bound > 0 and upper bound = 0).
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
deadEndMetabolites(mod)

}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{geneDel}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
