% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{elements}
\alias{elements}
\title{Data frame of elements and their average weights}
\usage{
elements()
}
\description{
Gets a data.frame with all elements, their symbol, name, atomic number and
their abridged standard atomic weight according to the International Union of
Pure and Applied Chemistry (IUPAC).
}
\references{
Prohaska T, Irrgeher J, Benefield J, Böhlke JK, Chesson LA, Coplen TB, et al.
Standard atomic weights of the elements 2021 (IUPAC Technical Report). Vol.
94, Pure and Applied Chemistry. 2022. p. 573–600.
http://dx.doi.org/10.1515/pac-2019-0603
}
