% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba.R
\name{fba}
\alias{fba}
\title{Flux Balance Analysis}
\usage{
fba(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\value{
A list with flux predictions (reaction fluxes 'fluxes', reduced costs 'redCosts'), and optimization status ()
}
\description{
Performs basic flux balance analysis (fba)
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)

# aerobic growth
res_aero <- fba(mod)
cat(" Growth rate:       ", res_aero@obj,"\n",
    "Acetate production:", res_aero@fluxes[mod@react_id == "EX_ac_e"],"\n")

mod <- changeBounds(mod, react = "EX_o2_e", lb = 0) # before: -1000
res_anaero <- fba(mod)
cat(" Growth rate:       ", res_anaero@obj,"\n",
    "Acetate production:", res_anaero@fluxes[mod@react_id == "EX_ac_e"],"\n")

}
\seealso{
Other Flux prediction algorithms:
\code{\link[=fva]{fva()}},
\code{\link[=pfba]{pfba()}},
\code{\link[=pfbaHeuristic]{pfbaHeuristic()}}
}
\concept{Flux prediction algorithms}
