% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findExchReact.R
\name{findExchReact}
\alias{findExchReact}
\title{Find exchange reactions}
\usage{
findExchReact(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg} or \link{ModelComm}}
}
\value{
If 'model' is of class \link{ModelOrg}, a data.frame is returned
stating all exchange reaction IDs, their index in the reaction list, the
respective metabolite id, name, and index in the metabolite list. If the
model is of class \link{ModelComm}, the resulting data.frame contains the
community exchange reactions and the organism-specific exchange reaction. The
latter are exchange reactions, that connect extracellular metabolites of the
organism metabolic network models with the shared extracellular space.
}
\description{
Finds all exchange reactions within a (community) model and report them in a
data frame.
}
