% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixBMRatios.R
\name{fixBMRatios}
\alias{fixBMRatios}
\title{Fixate biomass ratios}
\usage{
fixBMRatios(model, BMreact = guessBMReaction(model), tolerance = 0)
}
\arguments{
\item{model}{Community model of class \link{ModelComm}}

\item{BMreact}{IDs of individual biomass reactions}

\item{tolerance}{Tolerated deviation of ratios}
}
\description{
Biomass ratios are fixed via introducing new linear constraints.
}
\details{
Fixate the biomass ratios in a community models according to models' relative
abundances.
}
