% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxBMCoupling.R
\name{fluxBMCoupling}
\alias{fluxBMCoupling}
\title{Couple reaction flux bounds to biomass production}
\usage{
fluxBMCoupling(
  model,
  BMreact = guessBMReaction(model),
  cpl_c = 400,
  cpl_u = 0.01
)
}
\arguments{
\item{model}{Community model of class \link{ModelComm}}

\item{BMreact}{IDs of individual biomass reactions}

\item{cpl_c}{Coupling constraint \code{c}}

\item{cpl_u}{Coupling constraint \code{u}}
}
\description{
Resets the absolute flux bounds of reactions proportional to the flux through
the biomass reaction.
}
\details{
See (and cite) Heinken et al. (2013) Gut Microbes (doi: 10.4161/gmic.22370).
}
