% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frog.R
\docType{methods}
\name{frog}
\alias{frog}
\alias{frog,ModelOrg-method}
\alias{frog,ModelOrg}
\alias{frog,character-method}
\alias{frog,character}
\title{FROG: Reproducible fitness and robustness diagnostics for constraint-based
models}
\usage{
frog(model, ...)

\S4method{frog}{ModelOrg}(model, outdir = "frog_report")

\S4method{frog}{character}(model, outdir = "frog_report")
}
\arguments{
\item{model}{Model of class \link{ModelOrg} or path to SBML file.}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{outdir}{Path to output directory, where report files will be saved.}
}
\value{
TRUE if frog report export was successful.
}
\description{
Run a standardized FROG analysis on a model including (1) Objective Function
Values, (2) Flux Variability Analysis (FVA), (3) Gene Deletion Fluxes, (4)
Reaction Deletion Fluxes.
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)

output_dir <- tempdir()
frog(mod, outdir = output_dir)

# list files in output directory
dir(output_dir)

}
\references{
K. Raman et al., “FROG Analysis Ensures the Reproducibility of Genome Scale
Metabolic Models.,” bioRxiv, Sept. 2024. doi: 10.1101/2024.09.24.614797.
}
