% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{fvaJob}
\alias{fvaJob}
\alias{fvaJob,LPproblem_glpk-method}
\alias{fvaJob,LPproblem_glpk}
\title{Wrapper function for efficient FVA}
\usage{
fvaJob(lp, ...)

\S4method{fvaJob}{LPproblem_glpk}(lp, ind)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{ind}{Indices of variables to be tested in FVA}
}
\description{
Wrapper function for efficient FVA
}
\keyword{internal}
