% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{gene_num}
\alias{gene_num}
\alias{gene_num,ModelOrg-method}
\alias{gene_num,ModelOrg}
\title{Number of genes}
\usage{
gene_num(model)

\S4method{gene_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of genes of a model
}
