% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{gene_pos}
\alias{gene_pos}
\alias{gene_pos,modelorg,character-method}
\alias{gene_pos,modelorg,character}
\alias{gene_pos,modelorg,numeric-method}
\alias{gene_pos,modelorg,numeric}
\alias{gene_pos,modelorg,missing-method}
\alias{gene_pos,modelorg,missing}
\title{Index of gene(s)}
\usage{
gene_pos(model, gene)

\S4method{gene_pos}{modelorg,character}(model, gene)

\S4method{gene_pos}{modelorg,numeric}(model, gene)

\S4method{gene_pos}{modelorg,missing}(model, gene)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{gene}{Character vector with gene IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific gene(s).
}
\details{
Returns NA for gene IDs not part of the model or if the index is larger
than the number of genes in the model.
}
