% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{gene_pos}
\alias{gene_pos}
\title{Index of gene(s)}
\usage{
gene_pos(model, gene)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{gene}{Character vector with gene IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific gene(s).
}
\details{
Returns NA for gene IDs not part of the model or if the index is larger
than the number of gene in the model.
}
