% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDel.R
\name{geneDel}
\alias{geneDel}
\title{Identify reactions affected by gene knockouts}
\usage{
geneDel(model, gene)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{gene}{Character or numeric vector providing the IDs or indices of genes
to be deleted from 'model'.}
}
\value{
Character vector with reactions IDs.
}
\description{
This function identifies reactions, which cannot be catalyzed anymore when
a specified set of genes is deleted from a model.
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
