% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxdist_analysis.R
\name{getExchanges}
\alias{getExchanges}
\title{Get metabolite exchange rates}
\usage{
getExchanges(model, sol)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{sol}{Flux prediction results as an object of class
\link{FluxPrediction}.}
}
\value{
A data.frame that summarizes the predicted metabolite exchange rates
(=fluxes of exchange reactions).
}
\description{
Summarize the predicted fluxes for exchange reactions.
}
\seealso{
Other Flux analysis tools: 
\code{\link{FluxPrediction-class}},
\code{\link{show,FluxPrediction-method}}
}
\concept{Flux analysis tools}
