% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{getRedCosts}
\alias{getRedCosts}
\alias{getRedCosts,LPproblem_glpk-method}
\alias{getRedCosts,LPproblem_glpk}
\title{Retrieve column reduced costs}
\usage{
getRedCosts(lp, ...)

\S4method{getRedCosts}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Retrieve column reduced costs of a solved \link{LPproblem}.
}
\keyword{internal}
