% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessBMReaction.R
\docType{methods}
\name{guessBMReaction}
\alias{guessBMReaction}
\alias{guessBMReaction,ModelOrg-method}
\alias{guessBMReaction,ModelOrg}
\alias{guessBMReaction,ModelComm-method}
\alias{guessBMReaction,ModelComm}
\title{Guess biomass reaction(s)}
\usage{
guessBMReaction(model)

\S4method{guessBMReaction}{ModelOrg}(model)

\S4method{guessBMReaction}{ModelComm}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Guess biomass reaction(s)
}
