% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{loadMatrix}
\alias{loadMatrix}
\alias{loadMatrix,LPproblem_glpk-method}
\alias{loadMatrix,LPproblem_glpk}
\title{Populate a constraint-X-variable matrix}
\usage{
loadMatrix(lp, ...)

\S4method{loadMatrix}{LPproblem_glpk}(lp, ne, ia, ja, ra)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{ne}{Number of nonzero coefficients}

\item{ia}{row indices of nonzero coefficients}

\item{ja}{column indices of nonzero coefficients}

\item{ra}{nonzero values of respective entries}
}
\description{
Add linear coefficients to the constraint-X-variable matrix of an
\link{LPproblem}.
}
\keyword{internal}
