% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R
\docType{class}
\name{LPproblem-class}
\alias{LPproblem-class}
\alias{LPproblem}
\title{Structure of LPproblem Class}
\description{
A class structure to link LP problem C++ object.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to LP problem C++ object}

\item{\code{solver}}{Solver used for the LP problem}

\item{\code{method}}{Specific algorithm used by the LP solver}
}}

