% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblem_glpkClass.R
\docType{class}
\name{LPproblem_glpk-class}
\alias{LPproblem_glpk-class}
\title{Structure of LPproblem_glpk Class}
\description{
A class structure to link LP problem C++ object. Class is derived from
\link{LPproblem}
}
