% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoinformatics.R
\name{mass}
\alias{mass}
\title{Calculate molar mass of molecules}
\usage{
mass(formulas)
}
\arguments{
\item{formulas}{Character vector of chemical formulas.}
}
\description{
Calculates the average molar mass of compounds based on their chemical
formulas.
}
\examples{
mass(c("C6H12O6","C48H72CoN11O8","HCOOH"))

}
