% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{met_num}
\alias{met_num}
\alias{met_num,ModelOrg-method}
\alias{met_num,ModelOrg}
\title{Number of metabolites}
\usage{
met_num(model)

\S4method{met_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of metabolites of a model
}
