% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{met_num}
\alias{met_num}
\title{Number of metabolites}
\usage{
met_num(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of metabolites of a model
}
