% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{met_pos}
\alias{met_pos}
\title{Index of metabolite(s)}
\usage{
met_pos(model, met)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{met}{Character vector with metabolite IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific metabolite(s).
}
\details{
Returns NA for metabolite IDs not part of the model or if the index is larger
than the number of metabolites in the model.
}
