% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCommClass.R
\docType{class}
\name{ModelComm-class}
\alias{ModelComm-class}
\alias{ModelComm}
\title{Structure of ModelComm Class}
\description{
This class represents a community of model organisms.
}
\details{
Contains all slots from \link{ModelOrg-class}.
}
\section{Slots}{

\describe{
\item{\code{community}}{A data.frame}
}}

\seealso{
Other Object classes: 
\code{\link{Constraints-class}},
\code{\link{FluxPrediction-class}}
}
\concept{Object classes}
