% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{class}
\name{modelorg-class}
\alias{modelorg-class}
\alias{modelorg}
\title{Structure of modelorg Class}
\description{
This class represents a model organization with various attributes related
to central model structures, metabolites, reactions, and genes.
}
\section{Slots}{

\describe{
\item{\code{mod_id}}{A character vector representing the model identifier.}

\item{\code{mod_desc}}{A character vector describing the model.}

\item{\code{mod_name}}{A character vector containing the model name.}

\item{\code{mod_compart}}{A character vector indicating the model compartment.}

\item{\code{mod_compart_name}}{A character vector with the name of the model compartment.}

\item{\code{mod_attr}}{A data frame with additional model attributes.}

\item{\code{S}}{A sparse numeric matrix of \link[Matrix]{dgCMatrix-class} representing the Stoichiometric matrix.}

\item{\code{obj_coef}}{A numeric vector containing coefficients for the objective function.}

\item{\code{subSys}}{A sparse Boolean matrix of \link[Matrix]{lgCMatrix-class} defining subsystems.}

\item{\code{subSys_id}}{A character vector representing subsystem identifiers.}

\item{\code{subSys_name}}{A character vector containing the subsystem names.}

\item{\code{constraints}}{An object of class \link{Constraints} which specifies
constraints in a model in addition to stationarity and individual flux
bounds.}

\item{\code{met_id}}{A character vector representing metabolite identifiers.}

\item{\code{met_name}}{A character vector with metabolite names.}

\item{\code{met_comp}}{A character vector indicating metabolite compartments.}

\item{\code{met_attr}}{A data.frame that enables the storage of additional data for
metabolites. Only specific columns are exported to SBML files. See
\link{writeSBMLmod} for details.}

\item{\code{react_id}}{A character vector representing reaction identifiers.}

\item{\code{react_name}}{A character vector with reaction names.}

\item{\code{react_comp}}{A character vector indicating reaction compartments.}

\item{\code{lowbnd}}{A character vector containing lower bounds for reactions.}

\item{\code{uppbnd}}{A character vector containing upper bounds for reactions.}

\item{\code{react_attr}}{A data.frame that enables the storage of additional data for
reactions. Only specific columns are exported to SBML files. See
\link{writeSBMLmod} for details.}

\item{\code{gprRules}}{A character vector with Gene-Protein-Reaction association rules
(with gene product indices corresponding to the order in slot 'genes').}

\item{\code{genes}}{A list of character vectors. Each vector contains the IDs of
gene products associated to the respective reaction.}

\item{\code{allGenes}}{A character vector with all gene identifiers.}

\item{\code{genes_attr}}{A data.frame that enables the storage of additional data
(e.g., name and CVTerms) for genes/gene products. Only specific columns are
exported to SBML files. See \link{writeSBMLmod} for details.}
}}

