% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{printConstraint}
\alias{printConstraint}
\title{Print Constraint(s)}
\usage{
printConstraint(model, ind = NULL)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{ind}{Integer vector with the indices of the constraints to be printed}
}
\description{
Generate strings to summarize metabolic model constraints.
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{deadEndMetabolites}()},
\code{\link{geneDel}()},
\code{\link{printReaction}()}
}
\concept{Model characteristics}
