% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{printReaction}
\alias{printReaction}
\title{Print reaction(s)}
\usage{
printReaction(model, react, use.ids = FALSE)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{react}{A character vector specifying the reaction IDs or a integer
vector providing the reaction indices in the model.}

\item{use.ids}{Boolean. Indicating whether metabolite IDs should be printed
instead of metabolite names.}
}
\value{
A character vector with the individual reaction equations.
}
\description{
Print the equations of reactions.
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
# print reaction specified by index
printReaction(mod, react = 2)
# print reaction specified by ID
printReaction(mod, react = "PFL")
# print reaction with metabolite IDs instead of metabolite names
printReaction(mod, react = "PFL", use.ids = TRUE)
# print multiple reactions at once
printReaction(mod, react = c(2,8))

}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{geneDel}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
