% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{react_num}
\alias{react_num}
\alias{react_num,ModelOrg-method}
\alias{react_num,ModelOrg}
\title{Number of reactions}
\usage{
react_num(model)

\S4method{react_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of reactions of a model
}
\seealso{
Other Model characteristics: 
\code{\link{ModelOrg-class}},
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
