% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{react_num}
\alias{react_num}
\title{Number of reactions}
\usage{
react_num(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of reactions of a model
}
