% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{react_pos}
\alias{react_pos}
\title{Index of reaction(s)}
\usage{
react_pos(model, react)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{Character vector with reaction IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific reaction(s).
}
\details{
Returns NA for reaction IDs not part of the model or the index is larger than
the number of reactions in the model.
}
