% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbml_io.R
\name{readSBMLmod}
\alias{readSBMLmod}
\title{Reads an SBML file and constructs an object of class 'modelorg'}
\usage{
readSBMLmod(filename)
}
\arguments{
\item{filename}{Path to SBML file.}
}
\description{
Reads an SBML file and constructs an object of class 'modelorg'
}
