% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readSybilmod}
\alias{readSybilmod}
\title{Reads a sybil model file and constructs an object of cobrar's class
'ModelOrg'}
\usage{
readSybilmod(file_path)
}
\arguments{
\item{file_path}{Path to sybil model(s) saved as RDS file.}
}
\value{
A \link{ModelOrg-class} object or a list with \link{ModelOrg-class}
objects.
}
\description{
Reads a sybil model file and constructs an object of cobrar's class
'ModelOrg'
}
