% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{rmConstraint}
\alias{rmConstraint}
\title{Remove constraints}
\usage{
rmConstraint(model, ind)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{ind}{Integer vector with the indices of the constraints to be removed}
}
\description{
Remove specific user constraints from a metabolic model.
}
\seealso{
\code{\link[=printConstraint]{printConstraint()}}

Other Model manipulation tools: 
\code{\link{addCompartment}()},
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{addSubsystem}()},
\code{\link{changeBounds}()},
\code{\link{geneDel}()},
\code{\link{rmCompartment}()},
\code{\link{rmGene}()},
\code{\link{rmMetabolite}()},
\code{\link{rmReact}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
