% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{rmMetabolite}
\alias{rmMetabolite}
\title{Remove metabolites from a model}
\usage{
rmMetabolite(model, met)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{met}{A character vector stating the metabolite IDs in a model or a
numeric vector providing the metabolite indices.}
}
\value{
An updated model of class \link{modelorg}
}
\description{
This function removes specified metabolites from a model.
}
\note{
If at least one of the provided metabolites still participates in a reaction,
the function stops with an error message.
}
\seealso{
Other Model manipulation tools: 
\code{\link{addCompartment}()},
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{addSubsystem}()},
\code{\link{changeBounds}()},
\code{\link{geneDel}()},
\code{\link{rmCompartment}()},
\code{\link{rmConstraint}()},
\code{\link{rmGene}()},
\code{\link{rmReact}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
