% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{rmReact}
\alias{rmReact}
\title{Remove reactions from a model}
\usage{
rmReact(model, react, rm_met = TRUE)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{A character vector stating the reaction IDs in a model or a
numeric vector providing the reaction indices.}

\item{rm_met}{Logical. Should metabolites, which are singletons after the
reaction removal, be deleted as well?}
}
\value{
An updated model of class \link{modelorg}
}
\description{
This function removes specified reactions from a model.
}
\note{
If the reaction participates in a user constraint, this constraint is
removed from the model.
}
\seealso{
Other Model manipulation tools: 
\code{\link{addCompartment}()},
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{addSubsystem}()},
\code{\link{changeBounds}()},
\code{\link{geneDel}()},
\code{\link{rmCompartment}()},
\code{\link{rmConstraint}()},
\code{\link{rmGene}()},
\code{\link{rmMetabolite}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
