% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{setObjDir}
\alias{setObjDir}
\title{Set objective direction}
\usage{
setObjDir(model, dir)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{dir}{Character. Either "maximize" or "minimize".}
}
\value{
An updated model of class \link{ModelOrg}
}
\description{
Set the objective direction to either maximize or minimize
}
