% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxPredictionClass.R
\name{show,FluxPrediction-method}
\alias{show,FluxPrediction-method}
\title{Print a short summary of a flux prediction result}
\usage{
\S4method{show}{FluxPrediction}(object)
}
\arguments{
\item{object}{S4-object of class \link{FluxPrediction}.}
}
\description{
Displays a summary of the results obtained from a metabolic flux prediction
(e.g., FBA or pFBA).
}
