% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\name{show,ModelOrg-method}
\alias{show,ModelOrg-method}
\title{Print a short summary of a metabolic network model}
\usage{
\S4method{show}{ModelOrg}(object)
}
\arguments{
\item{object}{S4-object of class \link{ModelOrg}.}
}
\description{
Displays a few key properties of a metabolic network model of class
\link{ModelOrg}.
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{geneDel}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
