% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{subsys_num}
\alias{subsys_num}
\alias{subsys_num,modelorg-method}
\alias{subsys_num,modelorg}
\title{Number of subsystems}
\usage{
subsys_num(model)

\S4method{subsys_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of subsystems of a model
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{modelorg-class}},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
