% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{subsys_num}
\alias{subsys_num}
\title{Number of subsystems}
\usage{
subsys_num(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of subsystems of a model
}
