% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{subsys_pos}
\alias{subsys_pos}
\title{Index of subsystem(s)}
\usage{
subsys_pos(model, subsys)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{subsys}{Character vector with subsystem IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific subsystem(s).
}
\details{
Returns NA for subsystem IDs not part of the model or if the index is larger
than the number of subsystems in the model.
}
