% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{class}
\name{UserConstraints-class}
\alias{UserConstraints-class}
\alias{UserConstraints}
\title{Stucture of UserConstraints Class}
\description{
This class represents user constraints that can be added to a model of class
\link{modelorg} in addition to the stationarity constraint (\eqn{S v = 0})
and flux bounds.
}
\details{
The slot "rtype" describes the type of each constraint. Valid values and
their effects are:\tabular{clc}{
   \emph{code} \tab \emph{description} \tab \emph{rule} \cr
   "F" \tab free constraint \tab \eqn{-\infty < x < \infty} \cr
   "L" \tab constraint with lower bound \tab \eqn{lb \leq x \leq \infty} \cr
   "U" \tab constraint with upper bound \tab \eqn{-\infty \leq x \leq ub} \cr
   "D" \tab double-bounded (ranged) constraint \tab \eqn{lb \leq x \leq ub} \cr
   "E" \tab fixed (equality constraint) \tab \eqn{lb = x = ub} \cr
}
}
\section{Slots}{

\describe{
\item{\code{coeff}}{A sparse numeric matrix of \link[Matrix]{dgCMatrix-class}
representing the coefficients for each reaction in the model. Each row
denotes a user constraint, each column a reaction in the model in the same
order as in slor "S" in the corresponding \link{modelorg} object.}

\item{\code{lb}}{Numeric vector provididing the lower bound for each constraint.}

\item{\code{ub}}{Numeric vector provididing the lower bound for each constraint.}

\item{\code{rtype}}{Character vector stating the constraint type. See details.}
}}

