% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbml_io.R
\name{writeSBMLmod}
\alias{writeSBMLmod}
\title{Exports a Metabolic Network in SBML Format}
\usage{
writeSBMLmod(model, file_path = NULL)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{file_path}{SBML file name for exporting the model. Default is the
model's ID with ".xml" suffix.}
}
\value{
TRUE if file export was successful.
}
\description{
Export a constraint-based metabolic network model from a S4 object of class
\link{modelorg} to a SBML file.
}
\details{
SBML files are of level 3, version 2. FBC-package version 2.
}
