% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{addCols}
\alias{addCols}
\alias{addCols,LPproblem_glpk-method}
\alias{addCols,LPproblem_glpk}
\title{Add columns to LP problem}
\usage{
addCols(lp, ...)

\S4method{addCols}{LPproblem_glpk}(lp, ncols)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{ncols}{Number of columns to add}
}
\description{
Add columns (a.k.a. variables/reactions) to an \link{LPproblem}.
}
