% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addCompartment}
\alias{addCompartment}
\title{Add compartments or update their data}
\usage{
addCompartment(model, id, name = NA)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{id}{Character vector with compartment IDs}

\item{name}{Character vector for compartment names}
}
\description{
The function allows you to add one or more compartments to a model. When
providing the ID of an already existing compartment, you can use this
function to update the compartment's name.
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
mod <- addCompartment(mod, id = "p", name = "periplasm")

}
\seealso{
Other Model manipulation tools: 
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{addSubsystem}()},
\code{\link{changeBounds}()},
\code{\link{rmCompartment}()},
\code{\link{rmConstraint}()},
\code{\link{rmGene}()},
\code{\link{rmMetabolite}()},
\code{\link{rmReact}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
