% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addMetabolite}
\alias{addMetabolite}
\title{Add metabolites or update their data}
\usage{
addMetabolite(
  model,
  id,
  comp = NA,
  name = NA,
  chemicalFormula = NA,
  charge = NA,
  CVTerms = NA,
  SBOTerm = rep("SBO:0000247", length(id))
)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{id}{Character vector with metabolite IDs}

\item{comp}{Character vector of the metabolites' compartment IDs}

\item{name}{Character vector for metabolite names}

\item{chemicalFormula}{Character vector for the metabolites' chemical formulas}

\item{charge}{Numeric vector for the metabolites' charge}

\item{CVTerms}{Character vector for the metabolites' CV-Terms}

\item{SBOTerm}{Character vector for the metabolites' SBO-Terms}
}
\description{
The functions allows you to add one or more metabolites to a model. When
providing the ID of an already existing metabolite, you can use this function
to update metabolite information.
}
