% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{addSingleConstraint}
\alias{addSingleConstraint}
\alias{addSingleConstraint,LPproblem_glpk-method}
\alias{addSingleConstraint,LPproblem_glpk}
\title{Add single constraint}
\usage{
addSingleConstraint(lp, ...)

\S4method{addSingleConstraint}{LPproblem_glpk}(lp, coeffs, lb, ub, type)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{coeffs}{Linear coefficients for variables}

\item{lb}{Lower bound of constraint}

\item{ub}{Upper bound of constraint}

\item{type}{Constraint type}
}
\description{
Add a single constraint to an existing \link{LPproblem}.
}
