% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{addSubsystem}
\alias{addSubsystem}
\title{Add subsystems or update their data}
\usage{
addSubsystem(model, id, name = NA)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{id}{Character vector with subsystem IDs}

\item{name}{Character vector for subsystem names}
}
\description{
The function allows you to add one or more subsystems to a model. When
providing the ID of an already existing subsystem, you can use this
function to update the subsystem's name.
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
mod <- addSubsystem(mod, id = "Bifidoshunt",
                    name = "glucose fermentation to acetate and lactate (Bifidobacteria)")

}
\seealso{
Other Model manipulation tools: 
\code{\link{addCompartment}()},
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{changeBounds}()},
\code{\link{geneDel}()},
\code{\link{rmCompartment}()},
\code{\link{rmConstraint}()},
\code{\link{rmGene}()},
\code{\link{rmMetabolite}()},
\code{\link{rmReact}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
