% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkCompartmentId}
\alias{checkCompartmentId}
\title{Check compartment IDs and Indices}
\usage{
checkCompartmentId(model, comp)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{comp}{A character vector specifying the compartment IDs or a integer
vector providing the compartment indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether compartment IDs or indices are part of (valid) for a specific
model.
}
