% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkGeneId}
\alias{checkGeneId}
\title{Check gene IDs and Indices}
\usage{
checkGeneId(model, gene)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{gene}{A character vector specifying the gene IDs or a integer
vector providing the gene indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether gene IDs or indices are part of (valid) for a specific
model.
}
\seealso{
Other Model characteristics: 
\code{\link{ModelOrg-class}},
\code{\link{checkCompartmentId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
