% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkMetId}
\alias{checkMetId}
\title{Check metabolite IDs and Indices}
\usage{
checkMetId(model, met)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{A character vector specifying the metabolite IDs or a integer
vector providing the metabolite indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether metabolite IDs or indices are part of (valid) for a specific
model.
}
