% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReactId.R
\name{checkReactId}
\alias{checkReactId}
\title{Check reaction IDs and Indexes}
\usage{
checkReactId(model, react)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{A character vector specifying the reaction IDs or a integer
vector providing the reaction indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether reaction IDs or indexes are part of (valid) for a specific
model.
}
