% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkId.R
\name{checkSubsystemId}
\alias{checkSubsystemId}
\title{Check subsystem IDs and Indices}
\usage{
checkSubsystemId(model, subsystem)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{subsystem}{A character vector specifying the subsystem IDs or a integer
vector providing the subsystem indices in the model.}
}
\value{
A logical vector; TRUE if ID/index is valid, FALSE otherwise.
}
\description{
Checks whether subsystem IDs or indices are part of (valid) for a specific
model.
}
