% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{comp_num}
\alias{comp_num}
\alias{comp_num,modelorg-method}
\alias{comp_num,modelorg}
\title{Number of compartments}
\usage{
comp_num(model)

\S4method{comp_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of compartments of a model
}
