% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{comp_num}
\alias{comp_num}
\alias{comp_num,ModelOrg-method}
\alias{comp_num,ModelOrg}
\title{Number of compartments}
\usage{
comp_num(model)

\S4method{comp_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of compartments of a model
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{geneDel}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
