% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{comp_pos}
\alias{comp_pos}
\alias{comp_pos,ModelOrg,character-method}
\alias{comp_pos,ModelOrg,character}
\alias{comp_pos,ModelOrg,numeric-method}
\alias{comp_pos,ModelOrg,numeric}
\alias{comp_pos,ModelOrg,missing-method}
\alias{comp_pos,ModelOrg,missing}
\alias{comp_pos,ModelOrg,logical-method}
\alias{comp_pos,ModelOrg,logical}
\title{Index of compartment(s)}
\usage{
comp_pos(model, comp)

\S4method{comp_pos}{ModelOrg,character}(model, comp)

\S4method{comp_pos}{ModelOrg,numeric}(model, comp)

\S4method{comp_pos}{ModelOrg,missing}(model, comp)

\S4method{comp_pos}{ModelOrg,logical}(model, comp)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{comp}{Character vector with compartment IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific compartment(s).
}
\details{
Returns NA for compartment IDs not part of the model or if the index is larger
than the number of compartments in the model.
}
\seealso{
Other Model characteristics: 
\code{\link{ModelOrg-class}},
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{show,ModelOrg-method}},
\code{\link{subsys_num}()},
\code{\link{subsys_pos}()}
}
\concept{Model characteristics}
