% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{constraint_num}
\alias{constraint_num}
\alias{constraint_num,modelorg-method}
\alias{constraint_num,modelorg}
\title{Number of constraints}
\usage{
constraint_num(model)

\S4method{constraint_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of constraints of a model
}
