% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{deleteLP}
\alias{deleteLP}
\alias{deleteLP,LPproblem_glpk-method}
\alias{deleteLP,LPproblem_glpk}
\title{Delete an LP problem}
\usage{
deleteLP(lp, ...)

\S4method{deleteLP}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Deletes an existing \link{LPproblem} and frees associated memory.
}
