% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fva.R
\name{fva}
\alias{fva}
\title{Flux Variability Analysis (FVA)}
\usage{
fva(model, react = NULL, opt.factor = 1)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{react}{Character vector of reaction IDs tested for flux variability. If
NULL, all reactions are tested.}

\item{opt.factor}{Numeric value between 0 and 1 to define the required
fraction of the objective function value. E.g. 0.8 sets the constraint, that
in the flux variability analysis, the objective function value must at least
be 80\% of the original optimal value.}
}
\description{
Perform Flux Variability Analysis with or without relaxed optimality
constraint
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)

# Get flux variability for all exchange reactions
fvares <- fva(mod, react = mod@react_id[grepl("^EX_",mod@react_id)],
              opt.factor = 0.9)
fvares


}
