% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{gene_num}
\alias{gene_num}
\alias{gene_num,modelorg-method}
\alias{gene_num,modelorg}
\title{Number of genes}
\usage{
gene_num(model)

\S4method{gene_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of genes of a model
}
