% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDel.R
\name{geneDel}
\alias{geneDel}
\title{Identify reactions affected by gene knockouts}
\usage{
geneDel(model, gene)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{gene}{Character or numeric vector providing the IDs or indices of genes
to be deleted from 'model'.}
}
\value{
Character vector with reactions IDs.
}
\description{
This function identifies reactions, which cannot be catalyzed anymore when
a specified set of genes is deleted from a model.
}
\seealso{
Other Model manipulation tools: 
\code{\link{addCompartment}()},
\code{\link{addConstraint}()},
\code{\link{addGene}()},
\code{\link{addMetabolite}()},
\code{\link{addReact}()},
\code{\link{addSubsystem}()},
\code{\link{changeBounds}()},
\code{\link{rmCompartment}()},
\code{\link{rmConstraint}()},
\code{\link{rmGene}()},
\code{\link{rmMetabolite}()},
\code{\link{rmReact}()},
\code{\link{rmSubsystem}()}
}
\concept{Model manipulation tools}
