% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{getColsPrimal}
\alias{getColsPrimal}
\alias{getColsPrimal,LPproblem_glpk-method}
\alias{getColsPrimal,LPproblem_glpk}
\title{Retrieve column primal value}
\usage{
getColsPrimal(lp, ...)

\S4method{getColsPrimal}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Retrieve column primal value of a solved \link{LPproblem}.
}
