% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{getObjValue}
\alias{getObjValue}
\alias{getObjValue,LPproblem_glpk-method}
\alias{getObjValue,LPproblem_glpk}
\title{Get the objective value of a solved LP problem}
\usage{
getObjValue(lp, ...)

\S4method{getObjValue}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Get the objective value of a solved \link{LPproblem}.
}
