% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{getSolStat}
\alias{getSolStat}
\alias{getSolStat,LPproblem_glpk-method}
\alias{getSolStat,LPproblem_glpk}
\title{Get the solver status}
\usage{
getSolStat(lp, ...)

\S4method{getSolStat}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Get the solver status of a solved \link{LPproblem}.
}
