% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{met_num}
\alias{met_num}
\alias{met_num,modelorg-method}
\alias{met_num,modelorg}
\title{Number of metabolites}
\usage{
met_num(model)

\S4method{met_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of metabolites of a model
}
