% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{met_pos}
\alias{met_pos}
\alias{met_pos,ModelOrg,character-method}
\alias{met_pos,ModelOrg,character}
\alias{met_pos,ModelOrg,numeric-method}
\alias{met_pos,ModelOrg,numeric}
\alias{met_pos,ModelOrg,missing-method}
\alias{met_pos,ModelOrg,missing}
\title{Index of metabolite(s)}
\usage{
met_pos(model, met)

\S4method{met_pos}{ModelOrg,character}(model, met)

\S4method{met_pos}{ModelOrg,numeric}(model, met)

\S4method{met_pos}{ModelOrg,missing}(model, met)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{met}{Character vector with metabolite IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific metabolite(s).
}
\details{
Returns NA for metabolite IDs not part of the model or if the index is larger
than the number of metabolites in the model.
}
