% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCommClass.R
\docType{class}
\name{ModelComm-class}
\alias{ModelComm-class}
\alias{ModelComm}
\title{Structure of ModelComm Class}
\description{
This class represents a community of model organisms.
}
\section{Slots}{

\describe{
\item{\code{community}}{A data.frame}
}}

