% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfba.R
\name{pfba}
\alias{pfba}
\title{Parsimonious Flux Balance Analysis (pFBA)}
\usage{
pfba(model, costcoeffw = NULL, costcoefbw = NULL)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{costcoeffw, costcoefbw}{A numeric vector containing cost coefficients
for all variables/reactions (forward direction: 'costcoeffw'; backward
direction: 'costcoefbw'). If set to NULL, all cost coefficients are set to 1,
so that all variables have the same impact on the objective function.}
}
\value{
Returned reduced costs ('redCosts') correspond to the optimization of the
initial linear program (LP), which is basically the initial FBA to calculate
the optimal value of the objective function that is defined in 'model'.
}
\description{
Performs parsimonious FBA as describe by Lewis et al. 2010.
}
\references{
N. E. Lewis et al., “Omic data from evolved E. coli are consistent with
computed optimal growth from genome‐scale models,” Molecular Systems Biology,
vol. 6, no. 1. EMBO, Jan. 2010. doi: 10.1038/msb.2010.47.
}
\seealso{
\code{\link[=pfba_heuristic]{pfba_heuristic()}}
}
