% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printReaction.R
\name{printReaction}
\alias{printReaction}
\title{Print reaction(s)}
\usage{
printReaction(object, react, use.ids = FALSE)
}
\arguments{
\item{object}{Model of class \link{modelorg}}

\item{react}{A character vector specifying the reaction IDs or a integer
vector providing the reaction indices in the model.}

\item{use.ids}{Boolean. Indicating whether metabolite IDs should be printed
instead of metabolite names.}
}
\value{
A character vector with the individual reaction equations.
}
\description{
Print the equations of reactions.
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
# print reaction specified by index
printReaction(mod, react = 2)
# print reaction specified by ID
printReaction(mod, react = "PFL")
# print reaction with metabolite IDs instead of metabolite names
printReaction(mod, react = "PFL", use.ids = TRUE)
# print multiple reactions at once
printReaction(mod, react = c(2,8))

}
