% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{react_num}
\alias{react_num}
\alias{react_num,modelorg-method}
\alias{react_num,modelorg}
\title{Number of reactions}
\usage{
react_num(model)

\S4method{react_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of reactions of a model
}
