% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{react_num}
\alias{react_num}
\alias{react_num,ModelOrg-method}
\alias{react_num,ModelOrg}
\title{Number of reactions}
\usage{
react_num(model)

\S4method{react_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of reactions of a model
}
