% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{react_pos}
\alias{react_pos}
\alias{react_pos,ModelOrg,character-method}
\alias{react_pos,ModelOrg,character}
\alias{react_pos,ModelOrg,numeric-method}
\alias{react_pos,ModelOrg,numeric}
\alias{react_pos,ModelOrg,missing-method}
\alias{react_pos,ModelOrg,missing}
\title{Index of reaction(s)}
\usage{
react_pos(model, react)

\S4method{react_pos}{ModelOrg,character}(model, react)

\S4method{react_pos}{ModelOrg,numeric}(model, react)

\S4method{react_pos}{ModelOrg,missing}(model, react)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{react}{Character vector with reaction IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific reaction(s).
}
\details{
Returns NA for reaction IDs not part of the model or the index is larger than
the number of reactions in the model.
}
