% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbml_io.R
\name{readSBMLmod}
\alias{readSBMLmod}
\title{Reads an SBML file and constructs an object of class 'ModelOrg'}
\usage{
readSBMLmod(file_path)
}
\arguments{
\item{file_path}{Path to SBML file.}
}
\value{
A \link{ModelOrg-class} object.
}
\description{
Reads an SBML file and constructs an object of class 'ModelOrg'
}
\examples{
fpath <- system.file("extdata", "e_coli_core.xml", package="cobrar")
mod <- readSBMLmod(fpath)
mod

}
\seealso{
Other Model import and export helpers: 
\code{\link{readSybilmod}()},
\code{\link{writeSBMLmod}()}
}
\concept{Model import and export helpers}
