% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readsybilmod}
\alias{readsybilmod}
\title{Reads an sybil model file and constructs an object of class 'modelorg'}
\usage{
readsybilmod(file_path)
}
\arguments{
\item{file_path}{Path to sybil model(s) saved as RDS file.}
}
\value{
A \link{modelorg-class} object or a list with \link{modelorg-class} objects.
}
\description{
Reads an sybil model file and constructs an object of class 'modelorg'
}
