% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modifications.R
\name{rmConstraint}
\alias{rmConstraint}
\title{Remove constraints}
\usage{
rmConstraint(model, ind)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{ind}{Integer vector with the indices of the constraints to be removed}
}
\description{
Remove specific user constraints from a metabolic model.
}
\seealso{
\code{\link[=printConstraint]{printConstraint()}}
}
