% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{setColsBndsObjCoefs}
\alias{setColsBndsObjCoefs}
\alias{setColsBndsObjCoefs,LPproblem_glpk-method}
\alias{setColsBndsObjCoefs,LPproblem_glpk}
\title{Set column bounds and objective coefficients}
\usage{
setColsBndsObjCoefs(lp, ...)

\S4method{setColsBndsObjCoefs}{LPproblem_glpk}(lp, j, lb, ub, obj_coef, type = NULL)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{j}{Indices of variables to be updated}

\item{lb}{New lower bound of variable}

\item{ub}{New upper bound of variable}

\item{obj_coef}{New object coefficient of variable}

\item{type}{New type of column/variable}
}
\description{
Set column bounds and objective coefficients of an \link{LPproblem}.
}
