% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{setRowsBnds}
\alias{setRowsBnds}
\alias{setRowsBnds,LPproblem_glpk-method}
\alias{setRowsBnds,LPproblem_glpk}
\title{Set row bounds}
\usage{
setRowsBnds(lp, ...)

\S4method{setRowsBnds}{LPproblem_glpk}(lp, i, lb, ub, type)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}

\item{i}{Indices of rows/constraints}

\item{lb}{Lower bounds of constraints}

\item{ub}{Upper bound of constraints}

\item{type}{Type of constraint bounds}
}
\description{
Set row bounds of an \link{LPproblem}.
}
