% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\name{show,ModelOrg-method}
\alias{show,ModelOrg-method}
\title{Print a short summary of a metabolic network model}
\usage{
\S4method{show}{ModelOrg}(object)
}
\arguments{
\item{object}{S4-object of class \link{ModelOrg}.}
}
\description{
Displays a few key properties of a metabolic network model of class
\link{ModelOrg}.
}
