% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\name{show,modelorg-method}
\alias{show,modelorg-method}
\title{Print a short summary of a metabolic network model}
\usage{
\S4method{show}{modelorg}(object)
}
\arguments{
\item{object}{S4-object of class \link{modelorg}.}
}
\description{
Displays a few key properties of a metabolic network model of class
\link{modelorg}.
}
