% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPproblemClass.R, R/LPproblem_glpkClass.R
\docType{methods}
\name{solveLp}
\alias{solveLp}
\alias{solveLp,LPproblem_glpk-method}
\alias{solveLp,LPproblem_glpk}
\title{Solve an LP problem}
\usage{
solveLp(lp, ...)

\S4method{solveLp}{LPproblem_glpk}(lp)
}
\arguments{
\item{lp}{Object of class \link{LPproblem}}

\item{...}{Additional parameters passed on to the specific method instance.}
}
\description{
Solves an \link{LPproblem}.
}
