% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{subsys_num}
\alias{subsys_num}
\alias{subsys_num,ModelOrg-method}
\alias{subsys_num,ModelOrg}
\title{Number of subsystems}
\usage{
subsys_num(model)

\S4method{subsys_num}{ModelOrg}(model)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}
}
\description{
Get the total number of subsystems of a model
}
