% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{subsys_num}
\alias{subsys_num}
\alias{subsys_num,modelorg-method}
\alias{subsys_num,modelorg}
\title{Number of subsystems}
\usage{
subsys_num(model)

\S4method{subsys_num}{modelorg}(model)
}
\arguments{
\item{model}{Model of class \link{modelorg}}
}
\description{
Get the total number of subsystems of a model
}
