% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelorgClass.R
\docType{methods}
\name{subsys_pos}
\alias{subsys_pos}
\alias{subsys_pos,modelorg,character-method}
\alias{subsys_pos,modelorg,character}
\alias{subsys_pos,modelorg,numeric-method}
\alias{subsys_pos,modelorg,numeric}
\alias{subsys_pos,modelorg,missing-method}
\alias{subsys_pos,modelorg,missing}
\alias{subsys_pos,modelorg,logical-method}
\alias{subsys_pos,modelorg,logical}
\title{Index of subsystem(s)}
\usage{
subsys_pos(model, subsys)

\S4method{subsys_pos}{modelorg,character}(model, subsys)

\S4method{subsys_pos}{modelorg,numeric}(model, subsys)

\S4method{subsys_pos}{modelorg,missing}(model, subsys)

\S4method{subsys_pos}{modelorg,logical}(model, subsys)
}
\arguments{
\item{model}{Model of class \link{modelorg}}

\item{subsys}{Character vector with subsystem IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific subsystem(s).
}
\details{
Returns NA for subsystem IDs not part of the model or if the index is larger
than the number of subsystems in the model.
}
\seealso{
Other Model characteristics: 
\code{\link{checkCompartmentId}()},
\code{\link{checkGeneId}()},
\code{\link{checkMetId}()},
\code{\link{checkReactId}()},
\code{\link{checkSubsystemId}()},
\code{\link{comp_num}()},
\code{\link{comp_pos}()},
\code{\link{constraint_num}()},
\code{\link{deadEndMetabolites}()},
\code{\link{gene_num}()},
\code{\link{gene_pos}()},
\code{\link{met_num}()},
\code{\link{met_pos}()},
\code{\link{modelorg-class}},
\code{\link{printConstraint}()},
\code{\link{printReaction}()},
\code{\link{react_num}()},
\code{\link{react_pos}()},
\code{\link{subsys_num}()}
}
\concept{Model characteristics}
