% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOrgClass.R
\docType{methods}
\name{subsys_pos}
\alias{subsys_pos}
\alias{subsys_pos,ModelOrg,character-method}
\alias{subsys_pos,ModelOrg,character}
\alias{subsys_pos,ModelOrg,numeric-method}
\alias{subsys_pos,ModelOrg,numeric}
\alias{subsys_pos,ModelOrg,missing-method}
\alias{subsys_pos,ModelOrg,missing}
\alias{subsys_pos,ModelOrg,logical-method}
\alias{subsys_pos,ModelOrg,logical}
\title{Index of subsystem(s)}
\usage{
subsys_pos(model, subsys)

\S4method{subsys_pos}{ModelOrg,character}(model, subsys)

\S4method{subsys_pos}{ModelOrg,numeric}(model, subsys)

\S4method{subsys_pos}{ModelOrg,missing}(model, subsys)

\S4method{subsys_pos}{ModelOrg,logical}(model, subsys)
}
\arguments{
\item{model}{Model of class \link{ModelOrg}}

\item{subsys}{Character vector with subsystem IDs or Integer vector providing
indexes.}
}
\description{
Returns the index(es) of specific subsystem(s).
}
\details{
Returns NA for subsystem IDs not part of the model or if the index is larger
than the number of subsystems in the model.
}
